+++ src/genesys/roller.test.ts
@@ -14,28 +14,28 @@ test('should react to sw command', () => {
 });
 
 test('should roll various dice', () => {
-    const roller = genesysRoller(makeRng(1, 3, 5, 3, 4, 6, 7), 'gen');
+    const roller = genesysRoller(makeRng(5, 3, 5, 3, 1, 3, 7), 'gen');
     const result = roller.roll(new DicePool(1, 1, 1, 1, 1, 1, 1));
 
     expect(result.length).toBe(7);
-    expect(result[0].die).toBe(Dice.BOOST);
-    expect(result[0].face).toBe(Faces.BLANK);
+    expect(result[0].die).toBe(Dice.PROFICIENCY);
+    expect(result[0].face).toBe(Faces.ADVANTAGE);
     expect(result[1].die).toBe(Dice.ABILITY);
     expect(result[1].face).toBe(Faces.DOUBLE_SUCCESS);
-    expect(result[2].die).toBe(Dice.PROFICIENCY);
-    expect(result[2].face).toBe(Faces.ADVANTAGE);
-    expect(result[3].die).toBe(Dice.SETBACK);
-    expect(result[3].face).toBe(Faces.FAILURE);
-    expect(result[4].die).toBe(Dice.DIFFICULTY);
-    expect(result[4].face).toBe(Faces.THREAT);
-    expect(result[5].die).toBe(Dice.CHALLENGE);
-    expect(result[5].face).toBe(Faces.THREAT);
+    expect(result[2].die).toBe(Dice.CHALLENGE);
+    expect(result[2].face).toBe(Faces.THREAT);
+    expect(result[3].die).toBe(Dice.DIFFICULTY);
+    expect(result[3].face).toBe(Faces.THREAT);
+    expect(result[4].die).toBe(Dice.BOOST);
+    expect(result[4].face).toBe(Faces.BLANK);
+    expect(result[5].die).toBe(Dice.SETBACK);
+    expect(result[5].face).toBe(Faces.FAILURE);
     expect(result[6].die).toBe(Dice.FORCE);
     expect(result[6].face).toBe(Faces.FORCE);
 });
 
 test('should count results', () => {
-    const roller = genesysRoller(makeRng(1, 3, 5, 3, 4, 6, 7), 'gen');
+    const roller = genesysRoller(makeRng(5, 3, 5, 3, 1, 3, 7), 'gen');
     const result = roller.roll(new DicePool(1, 1, 1, 1, 1, 1, 1));
     const combined = combineRolls(result, parseRollValues, rollValuesMonoid);
 
@@ -51,7 +51,7 @@ test('should count results', () => {
 });
 
 test('should interpret results', () => {
-    const roller = genesysRoller(makeRng(1, 3, 5, 3, 4, 6, 7), 'gen');
+    const roller = genesysRoller(makeRng(5, 3, 5, 3, 1, 3, 7), 'gen');
     const result = roller.roll(new DicePool(1, 1, 1, 1, 1, 1, 1));
     const combined = combineRolls(result, parseRollValues, rollValuesMonoid);
     const interpreted = interpretResult(combined);