+++ src/genesys/roller.ts
@@ -39,12 +39,12 @@ export class GenesysRoller extends Roller<Dice, Faces, DicePool> {
 
     public roll(pool: DicePool): Array<Roll<Dice, Faces>> {
         return [
-            ...rollDie(pool.boost, Dice.BOOST, BOOST_ROLL_TABLE, this.rng),
-            ...rollDie(pool.ability, Dice.ABILITY, ABILITY_ROLL_TABLE, this.rng),
             ...rollDie(pool.proficiency, Dice.PROFICIENCY, PROFICIENCY_ROLL_TABLE, this.rng),
-            ...rollDie(pool.setback, Dice.SETBACK, SETBACK_ROLL_TABLE, this.rng),
-            ...rollDie(pool.difficulty, Dice.DIFFICULTY, DIFFICULTY_ROLL_TABLE, this.rng),
+            ...rollDie(pool.ability, Dice.ABILITY, ABILITY_ROLL_TABLE, this.rng),
             ...rollDie(pool.challenge, Dice.CHALLENGE, CHALLENGE_ROLL_TABLE, this.rng),
+            ...rollDie(pool.difficulty, Dice.DIFFICULTY, DIFFICULTY_ROLL_TABLE, this.rng),
+            ...rollDie(pool.boost, Dice.BOOST, BOOST_ROLL_TABLE, this.rng),
+            ...rollDie(pool.setback, Dice.SETBACK, SETBACK_ROLL_TABLE, this.rng),
             ...rollDie(pool.force, Dice.FORCE, FORCE_ROLL_TABLE, this.rng),
         ];
     }