+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.1] - 2020-05-18
+### Fixed
+- Increase die size from 25x25 to 40x40 pixels for better readability
+- Genesys dice are now show in the following order: proficiency, ability, challenge, difficulty, boost, setback, force
+
 ## [0.9.0] - 2020-05-18
 ### Changed
 - Make compatible with 0.5.7